Работает с переключением фильтров
